# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SaveSettingsResponse": {
        "description": "Acknowledgement returned after a successful settings save + reconnect.",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "message"
        ],
        "title": "SaveSettingsResponse",
        "type": "object"
      },
      "SettingsProgressResponse": {
        "description": "Current stage of the most recent (or in-flight) Save & Connect.",
        "properties": {
          "stage": {
            "description": "One of: idle, applying, capella_user, capella_bucket, connecting, creating_bucket, creating_collections, creating_indexes, building_search_index, saving, done, error.",
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "stage"
        ],
        "title": "SettingsProgressResponse",
        "type": "object"
      },
      "SettingsRequest": {
        "description": "Cluster + OpenAI settings submitted by the Settings UI.",
        "properties": {
          "azure_openai_endpoint": {
            "description": "Azure OpenAI resource URL (https://...).",
            "title": "Azure Openai Endpoint",
            "type": "string"
          },
          "capella_api_key_id": {
            "default": "",
            "description": "Capella Management API key ID (only when EMBEDDING_METHOD=capella).",
            "format": "password",
            "title": "Capella Api Key Id",
            "type": "string",
            "writeOnly": true
          },
          "capella_api_key_token": {
            "default": "",
            "description": "Capella Management API key token.",
            "format": "password",
            "title": "Capella Api Key Token",
            "type": "string",
            "writeOnly": true
          },
          "capella_workflow_name": {
            "default": "realtime_rag_vectorization",
            "description": "Capella AI Workflow name. Backend looks for an existing workflow with this name on the cluster and reuses it; creates one with this exact name if none exists.",
            "title": "Capella Workflow Name",
            "type": "string"
          },
          "cb_bucket": {
            "title": "Cb Bucket",
            "type": "string"
          },
          "cb_collection": {
            "description": "Collection name. UI auto-fills from ``embedding_method`` (documents_local / documents_capella) but the user can override.",
            "title": "Cb Collection",
            "type": "string"
          },
          "cb_connection_string": {
            "description": "``couchbases://...`` endpoint.",
            "title": "Cb Connection String",
            "type": "string"
          },
          "cb_password": {
            "format": "password",
            "title": "Cb Password",
            "type": "string",
            "writeOnly": true
          },
          "cb_scope": {
            "default": "_default",
            "title": "Cb Scope",
            "type": "string"
          },
          "cb_search_index": {
            "description": "Vector search index name. Auto-filled by UI based on mode.",
            "title": "Cb Search Index",
            "type": "string"
          },
          "cb_username": {
            "title": "Cb Username",
            "type": "string"
          },
          "deepgram_api_key": {
            "default": "",
            "description": "Deepgram primary STT API key (Whisper is the fallback when blank).",
            "format": "password",
            "title": "Deepgram Api Key",
            "type": "string",
            "writeOnly": true
          },
          "embedding_method": {
            "default": "capella",
            "description": "One of ``capella`` (AI Services workflow) or ``python`` (local).",
            "title": "Embedding Method",
            "type": "string"
          },
          "openai_api_key": {
            "description": "Azure OpenAI API key.",
            "format": "password",
            "title": "Openai Api Key",
            "type": "string",
            "writeOnly": true
          },
          "openai_embedding_model": {
            "description": "Azure OpenAI embedding deployment name.",
            "title": "Openai Embedding Model",
            "type": "string"
          },
          "openai_realtime_model": {
            "description": "Azure OpenAI Realtime deployment name.",
            "title": "Openai Realtime Model",
            "type": "string"
          },
          "tavily_api_key": {
            "default": "",
            "description": "Tavily key for the web-search fallback.",
            "format": "password",
            "title": "Tavily Api Key",
            "type": "string",
            "writeOnly": true
          },
          "web_search_enabled": {
            "default": false,
            "description": "Enable Tavily web-search fallback when KB has no relevant results.",
            "title": "Web Search Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "cb_connection_string",
          "cb_username",
          "cb_password",
          "cb_bucket",
          "cb_collection",
          "cb_search_index",
          "azure_openai_endpoint",
          "openai_api_key",
          "openai_realtime_model",
          "openai_embedding_model"
        ],
        "title": "SettingsRequest",
        "type": "object"
      },
      "SettingsResponse": {
        "description": "Envelope for :class:`SettingsValues`.",
        "properties": {
          "settings": {
            "$ref": "#/components/schemas/SettingsValues"
          }
        },
        "required": [
          "settings"
        ],
        "title": "SettingsResponse",
        "type": "object"
      },
      "SettingsStatusResponse": {
        "description": "Whether the backend currently has a live Couchbase connection.",
        "properties": {
          "initialized": {
            "title": "Initialized",
            "type": "boolean"
          }
        },
        "required": [
          "initialized"
        ],
        "title": "SettingsStatusResponse",
        "type": "object"
      },
      "SettingsValues": {
        "description": "Settings payload returned to the UI. Secret fields come back blank.",
        "properties": {
          "azure_openai_endpoint": {
            "title": "Azure Openai Endpoint",
            "type": "string"
          },
          "capella_api_key_id": {
            "title": "Capella Api Key Id",
            "type": "string"
          },
          "capella_api_key_token": {
            "title": "Capella Api Key Token",
            "type": "string"
          },
          "capella_workflow_name": {
            "title": "Capella Workflow Name",
            "type": "string"
          },
          "cb_bucket": {
            "title": "Cb Bucket",
            "type": "string"
          },
          "cb_collection": {
            "title": "Cb Collection",
            "type": "string"
          },
          "cb_connection_string": {
            "title": "Cb Connection String",
            "type": "string"
          },
          "cb_password": {
            "title": "Cb Password",
            "type": "string"
          },
          "cb_scope": {
            "title": "Cb Scope",
            "type": "string"
          },
          "cb_search_index": {
            "title": "Cb Search Index",
            "type": "string"
          },
          "cb_username": {
            "title": "Cb Username",
            "type": "string"
          },
          "deepgram_api_key": {
            "title": "Deepgram Api Key",
            "type": "string"
          },
          "embedding_method": {
            "title": "Embedding Method",
            "type": "string"
          },
          "openai_api_key": {
            "title": "Openai Api Key",
            "type": "string"
          },
          "openai_embedding_model": {
            "title": "Openai Embedding Model",
            "type": "string"
          },
          "openai_realtime_model": {
            "title": "Openai Realtime Model",
            "type": "string"
          },
          "tavily_api_key": {
            "title": "Tavily Api Key",
            "type": "string"
          },
          "web_search_enabled": {
            "title": "Web Search Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "cb_connection_string",
          "cb_username",
          "cb_password",
          "cb_bucket",
          "cb_scope",
          "cb_collection",
          "cb_search_index",
          "embedding_method",
          "azure_openai_endpoint",
          "openai_api_key",
          "openai_realtime_model",
          "openai_embedding_model",
          "capella_api_key_id",
          "capella_api_key_token",
          "capella_workflow_name",
          "deepgram_api_key",
          "tavily_api_key",
          "web_search_enabled"
        ],
        "title": "SettingsValues",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Couchbase Realtime Voice RAG",
      "url": "https://github.com/"
    },
    "description": "Reference backend for the Couchbase Realtime Voice RAG demo.\n\nThe API exposes authentication, document ingestion, chat-history persistence,\nCouchbase connection settings, and the WebSocket relay that bridges browser\naudio to the Azure OpenAI Realtime API. Document uploads are chunked and\neither embedded locally via ``text-embedding-3-small`` or vectorized through\nthe Capella AI Services workflow before being indexed for vector search.\n",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Couchbase Realtime Voice RAG",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/settings": {
      "get": {
        "description": "Return stored settings, including secret fields.\n\nAuthenticated callers get the same view they'd see in any password\nmanager: the secret values are echoed so the UI can pre-fill inputs\nthat render as ``type=password`` (dot-masked), with an eye toggle\nrevealing plaintext on demand. The on-disk store stays\nFernet-encrypted regardless.",
        "operationId": "get_settings_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read stored Couchbase settings",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "description": "Persist new connection settings, reconnect to Couchbase, and re-run index setup.\n\nIn Capella mode (and only when a Capella API key is configured), the\nDB user and bucket are auto-created via the Management API before the\nSDK connect attempt. The Docker scenario relies on\n``couchbase_service._init_docker_cluster`` to bootstrap services /\nquotas / admin \u2014 that's invoked from inside ``connect()``.",
        "operationId": "save_settings_api_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Couchbase settings and reconnect",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/progress": {
      "get": {
        "description": "Return the stage of the most recent (or in-flight) Save & Connect.\n\nPolled by the Settings UI on ~0.5s intervals while the form is\nsubmitting so the button label can advance through \"applying\",\n\"connecting\", \"creating_collections\", \"building_search_index\", etc.\nThe POST /api/settings call itself only resolves at the end of the\nmulti-step setup, so without this endpoint the user has no\nindication of which step is currently running.",
        "operationId": "settings_progress_api_settings_progress_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Current Save & Connect progress stage",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/status": {
      "get": {
        "description": "Return whether the server currently holds a live Couchbase connection.\n\nFalls back to the on-disk settings file if this worker's in-memory\n``cb_initialized`` flag is False. Multi-worker deployments only run\nlifespan-init once per process, so a request can land on a worker that\nmissed the original Save & Connect even though the cluster is fully\nset up. In that case we lazily apply the saved settings + reconnect on\nthe worker, mirroring what lifespan would have done.",
        "operationId": "settings_status_api_settings_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Check Couchbase connection status",
        "tags": [
          "settings"
        ]
      }
    }
  },
...
}